skills/discogs-sync/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discogs-sync"
version = "0.1.0"
description = "CLI tool to synchronize wantlists and collections with Discogs"
requires-python = ">=3.10"
dependencies = [
    "python3-discogs-client>=2.8",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "responses>=0.23"]

[project.scripts]
discogs-sync = "discogs_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
